requirements.txt
# MLOps 환경 필수 의존성

# 핵심 라이브러리
mlflow>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
joblib>=1.0.0

# 문서 처리
python-frontmatter>=1.0.0
markdown>=3.4.0
jinja2>=3.1.0
pyyaml>=6.0

# 파일 감시 및 스케줄링
watchdog>=2.1.0
schedule>=1.2.0

# HTTP 요청
requests>=2.28.0

# HuggingFace 통합
transformers>=4.21.0
datasets>=2.0.0
torch>=1.12.0

# 테스트
pytest>=7.0.0
pytest-cov>=4.0.0

# 로깅 및 유틸리티
python-dotenv>=0.19.0

# 개발 도구
black>=22.0.0
flake8>=5.0.0
mypy>=0.950
isort>=5.13.0

